requirements/base.txt
django==1.5
bpython==0.12
django-braces==0.2.1
django-model-utils==1.1.0
logutils==0.3.3
South==0.7.6
django-grappelli==2.4.5
django-taggit==0.10a1
